% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comb.R
\name{comb.expandGrid}
\alias{comb.expandGrid}
\title{comb.expandGrid}
\usage{
comb.expandGrid(data)
}
\arguments{
\item{data}{a list with items.}
}
\value{
a list containing the combinations of values of the data.
}
\description{
uses the 'expand.grid' function to generate combinations
}
