% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comb.R
\name{comb.randomExpandGrid}
\alias{comb.randomExpandGrid}
\title{comb.randomExpandGrid}
\usage{
comb.randomExpandGrid(data)
}
\arguments{
\item{data}{a list with items.}
}
\value{
a data.frame containing the combinations of values of the data.
}
\description{
uses the 'expand.grid' then will use the 'sample' function to randomise the order
to generate combinations
}
