% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{rFuncOpt_asDataFrame}
\alias{rFuncOpt_asDataFrame}
\title{rFuncOpt_asDataFrame}
\usage{
rFuncOpt_asDataFrame(data)
}
\arguments{
\item{data}{takes the return of function 'rFuncOpt'}
}
\value{
a data.frame
}
\description{
convert the result of rFuncOpt to a data.frame
}
