% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{rFuncOpt}
\alias{rFuncOpt}
\title{rFuncOpt}
\usage{
rFuncOpt(
  defaultFunction,
  params,
  combinationFunction = comb.expandGrid,
  combinationRuleFunction = NULL,
  thread.num = parallel::detectCores(),
  thread.type = ifelse(Sys.info()[["sysname"]] == "Windows", "PSOCK", "SOCK")
)
}
\arguments{
\item{defaultFunction}{a function that takes a lot of arguments ( should to use an elipsis "..." )
and return a list or vector containing the metrics you want.}

\item{params}{contains a list containing the name of parameter and its values.}

\item{combinationRuleFunction}{is a function to run for each combination in a list
remember 'the combination is a list' so each value of the list will be passed to this
function through of a loop' this function should to return the combination if wants to keep
or return NULL if the combination is unwanted.
it is usefull to remove unwanted combination for for example when a parameter
cant run ( is incompatible ) with another one.
this will run through foreach.}

\item{a}{function that takes the params and return a list of combinations. you can to use
any the functions from comb.*}
}
\value{
a list containing each run, this list is of the class 'rFuncOpt.result'
and can be passed to the function 'rFuncOpt_asDataFrame' to convert it to a data.frame.
}
\description{
does a optimization of function parameters through searching.
}
\examples{
# bellow is a sample of params
params<- list(first=1:5,second=5:10,third=list(tfirst=list('level2','level3','level4'),tsecond=list('level3','level4','level5')))
# bellow is a sample of a function, this function is returning three metrics for example.
defaultFunction<- function(...){ list(mae=2*rnorm(1),auc=3*rnorm(1),mape=5*rnorm(1)) }

# run this function
result<- rFuncOpt(defaultFunction=defaultFunction,params=params)
print.simple.list(result)

}
